add folders and move files into folders
cleanup project
