+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.CapstoneNewsApp"
         tools:targetApi="31">
         <activity
-            android:name=".NewsDetailActivity"
+            android:name=".views.NewsDetailActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
