+++ app/src/main/java/com/itscatalano/capstonenewsapp/models/APIresponse.kt
@@ -1,4 +1,4 @@
-package com.itscatalano.capstonenewsapp
+package com.itscatalano.capstonenewsapp.models
 
 /***
  * Author: Anthony Catalano
