+++ app/src/main/java/com/itscatalano/capstonenewsapp/views/NewsViewHolder.kt
@@ -1,13 +1,14 @@
-package com.itscatalano.capstonenewsapp
+package com.itscatalano.capstonenewsapp.views
 
 import androidx.recyclerview.widget.RecyclerView
+import com.itscatalano.capstonenewsapp.models.Article
 
 class NewsViewHolder (
     private val newsView : NewsView
 ): RecyclerView.ViewHolder(newsView){
 
 
-    fun bindData(article:Article, onTap: () -> Unit  ){
+    fun bindData(article: Article, onTap: () -> Unit  ){
         newsView.setData(article)
 
         newsView.setOnClickListener{
