+++ app/src/main/res/layout/activity_news_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NewsDetailActivity">
+    tools:context=".views.NewsDetailActivity">
 
     <TextView
         android:id="@+id/txtDescription"
